Monai/requirements.txt
torch>=1.4
pytorch-ignite==0.3.0
numpy
nibabel
tensorboard
pillow
scipy
scikit-image
monai==0.2.0

